Renamed and moved into service, fixed tests as well
