+++ app/src/androidTest/java/com/github/se/wanderpals/trip/MapScreenTests.kt
@@ -5,15 +5,14 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
 import androidx.test.platform.app.InstrumentationRegistry
-import com.github.se.wanderpals.BuildConfig
 import com.github.se.wanderpals.model.data.GeoCords
 import com.github.se.wanderpals.model.data.Stop
 import com.github.se.wanderpals.model.repository.TripsRepository
 import com.github.se.wanderpals.model.viewmodel.MapViewModel
 import com.github.se.wanderpals.screens.MapScreen
+import com.github.se.wanderpals.service.MapManager
 import com.github.se.wanderpals.ui.navigation.NavigationActions
 import com.github.se.wanderpals.ui.screens.trip.Map
-import com.google.android.libraries.places.api.Places
 import com.kaspersky.components.composesupport.config.withComposeSupport
 import com.kaspersky.kaspresso.kaspresso.Kaspresso
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
@@ -77,10 +76,10 @@ class MapScreenTests : TestCase(kaspressoBuilder = Kaspresso.Builder.withCompose
   @Before
   fun testSetup() {
     val context = InstrumentationRegistry.getInstrumentation().targetContext
-    Places.initialize(context, BuildConfig.MAPS_API_KEY)
-    val placesClient = Places.createClient(context)
+    val mapManager = MapManager(context)
+    mapManager.initClients()
     composeTestRule.setContent {
-      Map(mockNavActions, mapViewModel = FakeMapViewModel(), placesClient)
+      Map(mockNavActions, mapViewModel = FakeMapViewModel(), mapManager)
     }
   }
