+++ app/src/androidTest/java/com/github/se/wanderpals/trip/TripNavigationTest.kt
@@ -5,14 +5,13 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import com.github.se.wanderpals.BuildConfig
 import com.github.se.wanderpals.model.repository.TripsRepository
 import com.github.se.wanderpals.screens.TripScreen
+import com.github.se.wanderpals.service.MapManager
 import com.github.se.wanderpals.ui.navigation.NavigationActions
 import com.github.se.wanderpals.ui.navigation.Route
 import com.github.se.wanderpals.ui.navigation.rememberMultiNavigationAppState
 import com.github.se.wanderpals.ui.screens.trip.Trip
-import com.google.android.libraries.places.api.Places
 import com.kaspersky.components.composesupport.config.withComposeSupport
 import com.kaspersky.kaspresso.kaspresso.Kaspresso
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
@@ -51,9 +50,9 @@ class TripNavigationTest : TestCase(kaspressoBuilder = Kaspresso.Builder.withCom
                   rememberMultiNavigationAppState(
                       startDestination = Route.DASHBOARD, mockNavController2))
       val context = InstrumentationRegistry.getInstrumentation().targetContext
-      Places.initialize(context, BuildConfig.MAPS_API_KEY)
-      val placesClient = Places.createClient(context)
-      Trip(mockNavActions, "id", TripsRepository("-1", Dispatchers.IO), placesClient)
+      val mapManager = MapManager(context)
+      mapManager.initClients()
+      Trip(mockNavActions, "id", TripsRepository("-1", Dispatchers.IO), mapManager)
     }
   }
